Add a second rendering order based on node visit
- Move the layer methods to the scene node class
- Add new methods to the scene node class to rearrange children
- Add the new scene node class methods to the Lua API
- Delete the LayerBase enumeration from the drawable node class
- Use uint16_t instead of unsigned short for the layer variable
- Change the draw() method of a drawable node to return a boolean
- Modify the ImGui and Nuklear shader to use a depth value for vertices
- Add GUI widgets to show and manipulate layers in the anchor apptest

+++ src/scripting/LuaParticleSystem.cpp
@@ -569,7 +569,7 @@ int LuaParticleSystem::setLayer(lua_State *L)
 	ParticleSystem *particleSys = LuaClassWrapper<ParticleSystem>::unwrapUserData(L, -2);
 	const uint32_t layer = LuaUtils::retrieve<uint32_t>(L, -1);
 
-	particleSys->setLayer(layer);
+	particleSys->setLayer(static_cast<uint16_t>(layer));
 
 	return 0;
 }
